requirements.txt
faiss-cpu>=1.8.0
numpy>=1.26.0
sentence-transformers>=3.0.0
torch>=2.2.0

